+++ dotnet/src/UJMW.DynamicController/DynamicUjmwControllerFactory.DynamicControllerBase.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ApplicationParts;
 using Microsoft.AspNetCore.Mvc.Controllers;
@@ -6,12 +7,14 @@ using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Linq;
+using System.Net;
 using System.Reflection;
 using System.Reflection.Emit;
 using System.Reflection.PortableExecutable;
@@ -29,7 +32,6 @@ namespace System.Web.UJMW {
 
       private TServiceInterface _ServiceInstance;
 
-
       public DynamicControllerBase(TServiceInterface serviceInstance) {
         if (serviceInstance == null) {
           throw new Exception($"The dynamic asp controller base requires to get an '{typeof(TServiceInterface)}' injected - but got null!");
@@ -38,15 +40,24 @@ namespace System.Web.UJMW {
       }
 
       protected object InvokeMethod(string methodName, object requestDto) {
-        Func<TServiceInterface, object, object> redirector = GetOrCreateRedirectorMethod(methodName, this);
-        object responseDto = redirector.Invoke(_ServiceInstance, requestDto);
+
+        Func<TServiceInterface, object, IHeaderDictionary, IHeaderDictionary, object> redirector =
+          GetOrCreateRedirectorMethod(methodName, this);
+
+        object responseDto = redirector.Invoke(
+          _ServiceInstance,
+          requestDto,
+          this.HttpContext.Request.Headers,
+          this.HttpContext.Response.Headers
+        );        
+
         return responseDto;
       }
 
-      private static Dictionary<string, Func<TServiceInterface, object, object>> _RedirectorMethods =
-        new Dictionary<string, Func<TServiceInterface, object, object>>();
+      private static Dictionary<string, Func<TServiceInterface, object, IHeaderDictionary, IHeaderDictionary, object>> _RedirectorMethods =
+        new Dictionary<string, Func<TServiceInterface, object, IHeaderDictionary, IHeaderDictionary, object>>();
 
-      private static Func<TServiceInterface, object, object> GetOrCreateRedirectorMethod(string methodName, DynamicControllerBase<TServiceInterface> controller) {
+      private static Func<TServiceInterface, object, IHeaderDictionary, IHeaderDictionary, object> GetOrCreateRedirectorMethod(string methodName, DynamicControllerBase<TServiceInterface> controller) {
         lock (_RedirectorMethods) {
 
           if (_RedirectorMethods.ContainsKey(methodName)) {
@@ -88,8 +99,8 @@ namespace System.Web.UJMW {
           OutgoingResponseSideChannelConfiguration outboundSideChannelCfg = UjmwHostConfiguration.GetResponseSideChannelConfiguration(typeof(TServiceInterface));
 
           /////////// lambda (mth) ////////////////////////////////////////////
-          Func<TServiceInterface, object, object> mth = (
-            (svc, requestDto) => {
+          Func<TServiceInterface, object, IHeaderDictionary, IHeaderDictionary, object> mth = (
+            (svc, requestDto, requestHeaders, responseHeaders) => {
               object responseDto = Activator.CreateInstance(responseDtoType);
               object[] serviceMethodParams = new object[paramCount];
 
@@ -97,14 +108,55 @@ namespace System.Web.UJMW {
               foreach (DtoValueMapper requestDtoValueMapper in requestDtoValueMappers) {
                 requestDtoValueMapper.MapRequestDtoToParam(requestDto, serviceMethodParams);
               }
-
+          
               try {
 
-                //TODO: andere header auch auslesen bzw über alles iteriereun und am ende fallback
-                if (inboundSideChannelCfg.UnderlinePropertyIsAccepted && requestSidechannelProp != null) {
-                  var container = (Dictionary<string, string>)requestSidechannelProp.GetValue(requestDto);
-                  inboundSideChannelCfg.ProcessingMethod.Invoke(serviceMethod, container);
+                ///// RESTORE INCOMMING SIDECHANNEL /////
+
+                bool sideChannelReceived = false;
+                IDictionary<string, string> sideChannelContent = null;
+                foreach (string acceptedChannel in inboundSideChannelCfg.AcceptedChannels) {
+                  if (acceptedChannel == "_") {
+                    if (requestSidechannelProp != null) {
+                      var container = (Dictionary<string, string>)requestSidechannelProp.GetValue(requestDto);
+                      if (container != null) {
+                        sideChannelReceived = true;
+                        inboundSideChannelCfg.ProcessingMethod.Invoke(serviceMethod, container);
+                        break;
+                      }
+                    }
+                  }
+                  else { //lets look into the http header
+                    if (requestHeaders.ContainsKey(acceptedChannel)) {
+                      var rawSideChannelContent = requestHeaders[acceptedChannel].ToString();
+                      var serializer = new Newtonsoft.Json.JsonSerializer();
+                      sideChannelContent = JsonConvert.DeserializeObject<Dictionary<string, string>>(rawSideChannelContent);
+                      sideChannelReceived = true;
+                      inboundSideChannelCfg.ProcessingMethod.Invoke(serviceMethod, sideChannelContent);
+                      break;
+                    }
+                  }
                 }
+
+                if (!sideChannelReceived && inboundSideChannelCfg.AcceptedChannels.Length > 0) {
+                  if (inboundSideChannelCfg.SkipAllowed) {
+                    //if the whole getter is null, then (and only in this case) it will be a 'silent skip'
+                    if (inboundSideChannelCfg.DefaultsGetterOnSkip != null) {
+                      sideChannelContent = new Dictionary<string, string>();
+                      inboundSideChannelCfg.DefaultsGetterOnSkip.Invoke(ref sideChannelContent);
+                      //also null (when the DefaultsGetterOnSkip sets the ref handle to null) can be
+                      //passed to the processing method...
+                      inboundSideChannelCfg.ProcessingMethod.Invoke(serviceMethod, sideChannelContent);
+                    }
+                  }
+                  else {
+                    //TODO: logging hook!!! 
+                    Trace.TraceWarning("Rejected incomming request because of missing side channel");
+                    throw new Exception("Rejected incomming request because of missing side channel");
+                  }
+                }
+
+                ///// (end) RESTORE INCOMMING SIDECHANNEL /////
 
                 //invoke the service method
                 object returnVal = serviceMethod.Invoke(svc, serviceMethodParams);
@@ -130,20 +182,33 @@ namespace System.Web.UJMW {
                   faultProp.SetValue(responseDto, ex.Message);
                 }
               }
-              if (outboundSideChannelCfg.ChannelsToProvide.Any()) {
-                var backChannelContent = new Dictionary<string, string>();
 
-                ////// CAPTURE //////
-                outboundSideChannelCfg.CaptureMethod.Invoke(serviceMethod, backChannelContent);
-
-                if (outboundSideChannelCfg.UnderlinePropertyIsProvided && responseSidechannelProp != null) {
-                  responseSidechannelProp.SetValue(responseDto, backChannelContent);
-                }
+              ///// CAPTURE OUTGOING BACKCHANNEL /////
               
-                //TODO: die restlichen header auch noch
+              if (outboundSideChannelCfg.ChannelsToProvide.Any()) {
+                var backChannelContainer = new Dictionary<string, string>();
+                outboundSideChannelCfg.CaptureMethod.Invoke(serviceMethod, backChannelContainer);
+
+                //COLLECT THE DATA
+                string serializedSnapshot = null;
+                foreach (string channelName in outboundSideChannelCfg.ChannelsToProvide) {
+                  if (channelName == "_") {
+                    if (responseSidechannelProp != null) {
+                      responseSidechannelProp.SetValue(responseDto, backChannelContainer);
+                    }
+                  }
+                  else {
+                    if (serializedSnapshot == null) { //on-demand, but bufferred...
+                      serializedSnapshot = JsonConvert.SerializeObject(backChannelContainer);
+                    }
+                    responseHeaders.Add(channelName, serializedSnapshot);
+                  }
+                }
 
               }
-     
+
+              ///// (end) CAPTURE OUTGOING BACKCHANNEL /////
+
               return responseDto;
             }
           );/////// end of lambda (mth) ////////////////////////////////////////