+++ dotnet/test/UJMW.DemoWcfService/IDemoService.cs
@@ -1,5 +1,6 @@
 ﻿using DistributedDataFlow;
 using System;
+using System.IO;
 using System.Runtime.Serialization;
 using System.ServiceModel;
 using System.Threading;
@@ -15,8 +16,9 @@ namespace UJMW.DemoWcfService {
 #else
   [ServiceContract, SwaggerWcf("DemoService.svc"), SwaggerWcfServiceInfo("DemoService", "v1")]
 #endif
-  [HasDataFlowSideChannel("tenant-identifiers")]
-  public interface IDemoService : IDisposable{
+  [HasDataFlowSideChannel("tenant-identifiers"), HasDataFlowBackChannel("tenant-identifiers")]
+  public interface IDemoService : IDisposable {
+    
 
 #if NET5_0_OR_GREATER
 #else
@@ -29,6 +31,26 @@ namespace UJMW.DemoWcfService {
     [OperationContract, WebInvoke(Method = "POST")]
 #endif
     CompositeType GetDataUsingDataContract(CompositeType composite);
+
+  }
+
+  public interface IDemoFileService {
+
+    Stream Download1(string otp, out string fileName, out string fileContentType);
+    Stream Download2(string otp, out string fileContentType);
+    Stream Download3(string otp, out string fileName);
+    Stream Download4(string otp);
+
+    void Upload(string otp, Stream file, string fileName, string fileContentType);
+    void Upload(string otp, Stream file, string fileContentType);
+    void Upload(string otp, Stream file);
+    void Upload(string otp, Stream file1, string file1Name, string file1ContentType, Stream file2);
+    void Upload(string otp, Stream[] files, string[] fileNames, string[] fileContentTypes, Stream file2);
+
+    Stream UpAndDown(string otp,
+      Stream inputFile, string inputFileName, string inputFileContentType, 
+      out string fileName, out string fileContentType
+    );
 
   }